When bot starts it now checks if a guild(class) exists in Db and adds it if it doesnt.
To make this work properly i had to make changes to
Student,Teacher,Class models.
Their ID is not generated anymore so we have to manually add DiscordID when we shall add one of them to the DB.

Changes the seedmethod so the models gets an ID otherwise it crashed.